PriTesh
Save medical payment

+++ src/MedicalShopWeb/BusinessLayer/BLMedicalPayment.cs
@@ -22,7 +22,7 @@ namespace BusinessLayer
             return dsTotalAmt;
         }
 
-        public string SaveMedicalPayment(int SaleTransactionID, double PaidAmount, string PaymentDate, int UpdatedByUserID, string MedicalPaymentNo, double BalanceAmount,string coment)
+        public string SaveMedicalPayment(int SaleTransactionID, decimal PaidAmount, string PaymentDate, int UpdatedByUserID, string MedicalPaymentNo, decimal BalanceAmount,string coment)
         {
             string result = objMedicalPayment.SaveMedicalPayment(SaleTransactionID, PaidAmount, PaymentDate, UpdatedByUserID, MedicalPaymentNo, BalanceAmount,coment);
             return result;
